backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
google-generativeai==0.3.1
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
dj-database-url==2.1.0
